searchPage 이전

+++ dotori.go
@@ -40,7 +40,7 @@ var (
 	flagDBIP      = flag.String("dbip", "", "DB IP")
 	flagDBName    = flag.String("dbname", "dotori", "DB name")
 	flagHTTPPort  = flag.String("http", "", "Web Service Port Number")
-	flagPagenum   = flag.Int("pagenum", 9, "maximum number of items in a page")
+	flagPagenum   = flag.Int64("pagenum", 9, "maximum number of items in a page")
 	flagCookieAge = flag.Int("cookieage", 4, "cookie age (hour)") // MPAA 기준 4시간이다.
 
 	flagItemID = flag.String("itemid", "", "bson ObjectID assigned by mongodb")
